refactor(common): dedup AssetRef to single canonical declaration (ADR 0003)
Four copies of `AssetRef` collapse to one home at
`packages/common/src/types/asset-ref.ts`. The Zod schema is the
source of truth; the TS type derives via `z.infer`.

Removed:
- `state/assets/types.ts` — local `assetRefSchema` becomes a re-export.
- `state/requests/introspection.ts` — hand-written `interface AssetRef`
  (and its self-justifying duplicate-comment) becomes a re-export.
- `state/schemas/beak-project.ts` body.file — inline Zod replaced with
  the canonical `assetRefSchema`.
- `state/schemas/request-values.ts` bodyValue.file — same.

The SDK (`@getbeak/extension-sdk`) carries a structurally-mirrored
interface for extension authors; the comment in
`common/types/asset-ref.ts` notes the pair should stay in sync.

All workspaces typecheck clean. `@beak/state` 625/625 tests pass.
